refactor: MemberService 인터페이스 제거
